src/requirements.txt
pandas
numpy
scikit-learn
nltk
fastapi
uvicorn[standard]
joblib
aiohttp
kfp.v2
kfp.v2.dsl
# spaCy is more robust for NER, but the user's code uses NLTK's NE chunker.
# We will stick to the user's NLTK approach for consistency, but keep the option open.
# If NLTK NE chunking proves insufficient, spaCy will be added.
